passport-photo-ai/backend_backup_perfect_cropping_20251226_111332/requirements-full.txt
flask
flask_cors
Pillow
opencv-python-headless
numpy
python-dotenv
pillow-heif
rembg
gunicorn
anthropic
boto3
pytest
hypothesis
pytest-mock
pytest-cov